fix(djinn-db): keep queries.rs under the size guard
Adding the `merged` pseudo-status pushed queries.rs past the 50KB Server
Size Guard limit (it was only 219 bytes under). Extract the self-contained
`ready_projection_tests` module into its own child file
(queries/ready_projection_tests.rs) — a zero-logic-change split that
restores ~9KB of headroom — and keep the full `taskToColumnKey` sync
comment on the merged predicate.
